split the enmu as file
