+++ .swiftlint.yml
@@ -3,7 +3,6 @@ disabled_rules: # rule identifiers to exclude from running
   # - comma
   - control_statement
   - trailing_whitespace
-  - file_length
 opt_in_rules: # some rules are only opt-in
   - empty_count
   # Find all the available rules by running:
