+++ ZVProgressHUD/AnimationType.swift
@@ -1,0 +1,14 @@
+//
+//  AnimationType.swift
+//  ZVProgressHUD
+//
+//  Created by 张伟 on 2019/10/8.
+//  Copyright © 2019 zevwings. All rights reserved.
+//
+
+import UIKit
+
+public enum AnimationType {
+    case flat
+    case native
+}
