+++ ZVProgressHUD/Position.swift
@@ -1,0 +1,15 @@
+//
+//  Position.swift
+//  ZVProgressHUD
+//
+//  Created by 张伟 on 2019/10/8.
+//  Copyright © 2019 zevwings. All rights reserved.
+//
+
+import UIKit
+
+public enum Position {
+    case top
+    case center
+    case bottom
+}
